requirements.txt
pyadb
pyserial
RPIO
